+++ src/cpptools.ts
@@ -14,7 +14,7 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import * as cpt from 'vscode-cpptools';
 import * as nls from 'vscode-nls';
-import { TargetTypeString } from './drivers/cms-client';
+import { TargetTypeString } from './drivers/cmakeServerClient';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
