+++ src/drivers/cmakeServerApiDriver.ts
@@ -7,9 +7,9 @@ import * as vscode from 'vscode';
 import * as api from '@cmt/api';
 import { CacheEntryProperties, ExecutableTarget, RichTarget } from '@cmt/api';
 import * as cache from '@cmt/cache';
-import * as cms from '@cmt/drivers/cms-client';
+import * as cms from '@cmt/drivers/cmakeServerClient';
 import * as codemodel from '@cmt/drivers/codemodel-driver-interface';
-import { CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/driver';
+import { CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/cmakeDriver';
 import { Kit, CMakeGenerator } from '@cmt/kit';
 import { createLogger } from '@cmt/logging';
 import * as proc from '@cmt/proc';
@@ -29,7 +29,7 @@ export class NoGeneratorError extends Error {
     message: string = localize('no.usable.generator.found', 'No usable generator found.');
 }
 
-export class CMakeServerClientDriver extends CMakeDriver {
+export class CMakeServerApiDriver extends CMakeDriver {
 
     get isCacheConfigSupported(): boolean {
         return false;
@@ -411,8 +411,8 @@ export class CMakeServerClientDriver extends CMakeDriver {
         testPreset: TestPreset | null,
         workspaceFolder: string | null,
         preconditionHandler: CMakePreconditionProblemSolver,
-        preferredGenerators: CMakeGenerator[]): Promise<CMakeServerClientDriver> {
-        return this.createDerived(new CMakeServerClientDriver(cmake, config, workspaceFolder, preconditionHandler),
+        preferredGenerators: CMakeGenerator[]): Promise<CMakeServerApiDriver> {
+        return this.createDerived(new CMakeServerApiDriver(cmake, config, workspaceFolder, preconditionHandler),
             useCMakePresets,
             kit,
             configurePreset,
