+++ test/unit-tests/driver/driver-test.ts
@@ -7,13 +7,13 @@ import { expect } from 'chai';
 import * as chaiString from 'chai-string';
 import * as fs from 'fs';
 import * as path from 'path';
-import { CMakeFileApiDriver } from '@cmt/drivers/cmfileapi-driver';
-import { CMakeServerClientDriver, NoGeneratorError } from '@cmt/drivers/cms-driver';
+import { CMakeFileApiDriver } from '@cmt/drivers/cmakeFileApiDriver';
+import { CMakeServerApiDriver, NoGeneratorError } from '@cmt/drivers/cmakeServerApiDriver';
 
 chai.use(chaiString);
 
 import { Kit, CMakeGenerator } from '@cmt/kit';
-import { CMakePreconditionProblems, CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/driver';
+import { CMakePreconditionProblems, CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/cmakeDriver';
 
 const here = __dirname;
 function getTestRootFilePath(filename: string): string {
@@ -358,7 +358,7 @@ export function makeDriverTestsuite(driverName: string, driver_generator: (cmake
             expect(await driver.configure(ConfigureTrigger.runTests, [])).to.be.eq(0);
 
             const expFileApi = driver instanceof CMakeFileApiDriver;
-            const expSrv = driver instanceof CMakeServerClientDriver;
+            const expSrv = driver instanceof CMakeServerApiDriver;
             expect(!expFileApi || !expSrv); // mutually exclusive
 
             // Configure with a different generator should overwrite the previous Ninja generator
